django-api/requirements.txt
django
djangorestframework
gunicorn
whitenoise